server/requirements.txt
bottle
psycopg2
PyJWT
meinheld
gunicorn
